groinc 0.0.9

+++ network/my_types.h
@@ -13,8 +13,8 @@ typedef unsigned char __u8;
 typedef signed short __s16;
 typedef unsigned short __u16;
 
-typedef signed int __s32;
-typedef unsigned int __u32;
+typedef signed long int __s32;
+typedef unsigned long int __u32;
 
 typedef signed long long __s64;
 typedef unsigned long long __u64;
